requirements.txt
streamlit==1.41.1
openai
python-dotenv